WIP move ellipsis menu to the top of reader card (#66749)
