+++ client/blocks/reader-post-card/compact.jsx
@@ -1,6 +1,5 @@
 import PropTypes from 'prop-types';
 import ReaderExcerpt from 'calypso/blocks/reader-excerpt';
-import ReaderPostOptionsMenu from 'calypso/blocks/reader-post-options-menu';
 import AutoDirection from 'calypso/components/auto-direction';
 import FeaturedAsset from './featured-asset';
 
@@ -15,12 +14,6 @@ const CompactPost = ( { post, postByline, children, isDiscover, onClick } ) => {
 			/>
 			<div className="reader-post-card__post-details">
 				{ postByline }
-				<ReaderPostOptionsMenu
-					className="ignore-click"
-					showFollow={ true }
-					post={ post }
-					position="bottom"
-				/>
 				<AutoDirection>
 					<h2 className="reader-post-card__title">
 						<a className="reader-post-card__title-link" href={ post.URL }>
